tool, commandoutput: move tool-call context into the tool package
The loop and dispatcher stamped per-call metadata via
commandoutput.WithCallContext, coupling core to a concrete component
package. CallContext is a tool-invocation concern — the neutral tool
package the loop already legitimately imports — so the type and its
context helpers move there verbatim. dispatch.go no longer imports
commandoutput at all; loop.go keeps only the Store field pending its
interface extraction.
