+++ harness/internal/tool/builtins/shell.go
@@ -240,7 +240,7 @@ func ReadCommandOutputTool(store *commandoutput.Store, execs ...executor.Executo
 				return tool.StructuredResult{}, err
 			}
 			text := string(textBytes)
-			meta := commandoutput.CallContextFrom(ctx)
+			meta := tool.CallContextFrom(ctx)
 			if err := store.RecordRead(meta, params.Ref, read, text); err != nil {
 				return tool.StructuredResult{}, fmt.Errorf("record command output read: %w", err)
 			}
